Revert "seperate tree item from enitty"
